Combine FFBS functions into a single file. Create array for forward probabilities outside of the forward algorithm function.
