+++ R/RcppExports.R
@@ -1,21 +1,6 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-#' Backward Sampler for the Bounded Coalescent
-#'
-#' Samples the number of lineages at specified times given forward
-#' probabilities
-#'
-#' @param forward_probs 2D array of probabilities from the forward algorithm.
-#' @param times Vector of ordered sampling times for leaves.
-#' @param leaves Number of leaves taken at each sampling time.
-#' @param ne Effective population size.
-#' @param bound Bound time.
-#' @param bound_size Number of lineages at the bound (default 1).
-backward_sampler_c <- function(forward_probs, times, leaves, ne, bound, bound_size = 1L) {
-    .Call(`_BoundedCoalescent_backward_sampler_c`, forward_probs, times, leaves, ne, bound, bound_size)
-}
-
 #' Determine blocks of coalescent events from backward sample
 #'
 #' @param sample Sample of number of lineages from backward sampler
@@ -58,29 +43,6 @@ sample_coalescence_time_c <- function(time_lower, time_upper, lineages, ne) {
     .Call(`_BoundedCoalescent_sample_coalescence_time_c`, time_lower, time_upper, lineages, ne)
 }
 
-#' Forward Algorithm for the Bounded Coalescent
-#'
-#' Calculate the forward probabilities for the bounded coalescent.
-#'
-#' @param times Vector of ordered sampling times for leaves.
-#' @param leaves Number of leaves taken at each sampling time.
-#' @param ne Effective population size.
-#' @param bound Bound time.
-forward_algorithm_c <- function(times, leaves, ne, bound) {
-    .Call(`_BoundedCoalescent_forward_algorithm_c`, times, leaves, ne, bound)
-}
-
-#' Sample coalescence times for the bounded coalescenct
-#'
-#' @param times Vector of ordered sampling times for leaves.
-#' @param leaves Number of leaves taken at each sampling time.
-#' @param ne Effective population size.
-#' @param bound Bound time.
-#' @param nsam Number of samples (deafault 1).
-sample_bounded_times_c <- function(times, leaves, ne, bound, nsam = 1L) {
-    .Call(`_BoundedCoalescent_sample_bounded_times_c`, times, leaves, ne, bound, nsam)
-}
-
 #' Calculate coalescent probabilities (homochronous)
 #'
 #' Calculate the probability of i lineages coalescing down to j <= i lineages
@@ -92,6 +54,45 @@ sample_bounded_times_c <- function(times, leaves, ne, bound, nsam = 1L) {
 #' @param ne Effective population size.
 homochronous_probability <- function(i, j, dt, ne) {
     .Call(`_BoundedCoalescent_homochronous_probability`, i, j, dt, ne)
+}
+
+#' Forward Algorithm for the Bounded Coalescent
+#'
+#' Calculate the forward probabilities for the bounded coalescent.
+#'
+#' @param times Vector of ordered sampling times for leaves.
+#' @param leaves Number of leaves taken at each sampling time.
+#' @param ne Effective population size.
+#' @param bound Bound time.
+#' @param forward_probs Array to store probabilities.
+forward_algorithm_c <- function(times, leaves, ne, bound, forward_probs) {
+    invisible(.Call(`_BoundedCoalescent_forward_algorithm_c`, times, leaves, ne, bound, forward_probs))
+}
+
+#' Backward Sampler for the Bounded Coalescent
+#'
+#' Samples the number of lineages at specified times given forward
+#' probabilities
+#'
+#' @param forward_probs 2D array of probabilities from the forward algorithm.
+#' @param times Vector of ordered sampling times for leaves.
+#' @param leaves Number of leaves taken at each sampling time.
+#' @param ne Effective population size.
+#' @param bound Bound time.
+#' @param bound_size Number of lineages at the bound (default 1).
+backward_sampler_c <- function(forward_probs, times, leaves, ne, bound, bound_size = 1L) {
+    .Call(`_BoundedCoalescent_backward_sampler_c`, forward_probs, times, leaves, ne, bound, bound_size)
+}
+
+#' Sample coalescence times for the bounded coalescenct
+#'
+#' @param times Vector of ordered sampling times for leaves.
+#' @param leaves Number of leaves taken at each sampling time.
+#' @param ne Effective population size.
+#' @param bound Bound time.
+#' @param nsam Number of samples (deafault 1).
+sample_bounded_times_c <- function(times, leaves, ne, bound, nsam = 1L) {
+    .Call(`_BoundedCoalescent_sample_bounded_times_c`, times, leaves, ne, bound, nsam)
 }
 
 #' Calculate the likelihood under the bounded coalescence
