+++ src/RcppExports.cpp
@@ -10,22 +10,6 @@ Rcpp::Rostream<true>&  Rcpp::Rcout = Rcpp::Rcpp_cout_get();
 Rcpp::Rostream<false>& Rcpp::Rcerr = Rcpp::Rcpp_cerr_get();
 #endif
 
-// backward_sampler_c
-Rcpp::List backward_sampler_c(Rcpp::NumericVector forward_probs, Rcpp::NumericVector times, Rcpp::IntegerVector leaves, double ne, double bound, int bound_size);
-RcppExport SEXP _BoundedCoalescent_backward_sampler_c(SEXP forward_probsSEXP, SEXP timesSEXP, SEXP leavesSEXP, SEXP neSEXP, SEXP boundSEXP, SEXP bound_sizeSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type forward_probs(forward_probsSEXP);
-    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type times(timesSEXP);
-    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type leaves(leavesSEXP);
-    Rcpp::traits::input_parameter< double >::type ne(neSEXP);
-    Rcpp::traits::input_parameter< double >::type bound(boundSEXP);
-    Rcpp::traits::input_parameter< int >::type bound_size(bound_sizeSEXP);
-    rcpp_result_gen = Rcpp::wrap(backward_sampler_c(forward_probs, times, leaves, ne, bound, bound_size));
-    return rcpp_result_gen;
-END_RCPP
-}
 // block_coalescences_c
 Rcpp::DataFrame block_coalescences_c(Rcpp::IntegerVector sample, Rcpp::NumericVector times, Rcpp::IntegerVector leaves, double bound);
 RcppExport SEXP _BoundedCoalescent_block_coalescences_c(SEXP sampleSEXP, SEXP timesSEXP, SEXP leavesSEXP, SEXP boundSEXP) {
@@ -84,17 +68,47 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
-// forward_algorithm_c
-Rcpp::NumericVector forward_algorithm_c(Rcpp::NumericVector times, Rcpp::IntegerVector leaves, double ne, double bound);
-RcppExport SEXP _BoundedCoalescent_forward_algorithm_c(SEXP timesSEXP, SEXP leavesSEXP, SEXP neSEXP, SEXP boundSEXP) {
+// homochronous_probability
+double homochronous_probability(int i, int j, double dt, double ne);
+RcppExport SEXP _BoundedCoalescent_homochronous_probability(SEXP iSEXP, SEXP jSEXP, SEXP dtSEXP, SEXP neSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< int >::type i(iSEXP);
+    Rcpp::traits::input_parameter< int >::type j(jSEXP);
+    Rcpp::traits::input_parameter< double >::type dt(dtSEXP);
+    Rcpp::traits::input_parameter< double >::type ne(neSEXP);
+    rcpp_result_gen = Rcpp::wrap(homochronous_probability(i, j, dt, ne));
+    return rcpp_result_gen;
+END_RCPP
+}
+// forward_algorithm_c
+void forward_algorithm_c(Rcpp::NumericVector times, Rcpp::IntegerVector leaves, double ne, double bound, Rcpp::NumericVector forward_probs);
+RcppExport SEXP _BoundedCoalescent_forward_algorithm_c(SEXP timesSEXP, SEXP leavesSEXP, SEXP neSEXP, SEXP boundSEXP, SEXP forward_probsSEXP) {
+BEGIN_RCPP
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< Rcpp::NumericVector >::type times(timesSEXP);
     Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type leaves(leavesSEXP);
     Rcpp::traits::input_parameter< double >::type ne(neSEXP);
     Rcpp::traits::input_parameter< double >::type bound(boundSEXP);
-    rcpp_result_gen = Rcpp::wrap(forward_algorithm_c(times, leaves, ne, bound));
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type forward_probs(forward_probsSEXP);
+    forward_algorithm_c(times, leaves, ne, bound, forward_probs);
+    return R_NilValue;
+END_RCPP
+}
+// backward_sampler_c
+Rcpp::List backward_sampler_c(Rcpp::NumericVector forward_probs, Rcpp::NumericVector times, Rcpp::IntegerVector leaves, double ne, double bound, int bound_size);
+RcppExport SEXP _BoundedCoalescent_backward_sampler_c(SEXP forward_probsSEXP, SEXP timesSEXP, SEXP leavesSEXP, SEXP neSEXP, SEXP boundSEXP, SEXP bound_sizeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type forward_probs(forward_probsSEXP);
+    Rcpp::traits::input_parameter< Rcpp::NumericVector >::type times(timesSEXP);
+    Rcpp::traits::input_parameter< Rcpp::IntegerVector >::type leaves(leavesSEXP);
+    Rcpp::traits::input_parameter< double >::type ne(neSEXP);
+    Rcpp::traits::input_parameter< double >::type bound(boundSEXP);
+    Rcpp::traits::input_parameter< int >::type bound_size(bound_sizeSEXP);
+    rcpp_result_gen = Rcpp::wrap(backward_sampler_c(forward_probs, times, leaves, ne, bound, bound_size));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -110,20 +124,6 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< double >::type bound(boundSEXP);
     Rcpp::traits::input_parameter< int >::type nsam(nsamSEXP);
     rcpp_result_gen = Rcpp::wrap(sample_bounded_times_c(times, leaves, ne, bound, nsam));
-    return rcpp_result_gen;
-END_RCPP
-}
-// homochronous_probability
-double homochronous_probability(int i, int j, double dt, double ne);
-RcppExport SEXP _BoundedCoalescent_homochronous_probability(SEXP iSEXP, SEXP jSEXP, SEXP dtSEXP, SEXP neSEXP) {
-BEGIN_RCPP
-    Rcpp::RObject rcpp_result_gen;
-    Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< int >::type i(iSEXP);
-    Rcpp::traits::input_parameter< int >::type j(jSEXP);
-    Rcpp::traits::input_parameter< double >::type dt(dtSEXP);
-    Rcpp::traits::input_parameter< double >::type ne(neSEXP);
-    rcpp_result_gen = Rcpp::wrap(homochronous_probability(i, j, dt, ne));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -172,14 +172,14 @@ END_RCPP
 }
 
 static const R_CallMethodDef CallEntries[] = {
-    {"_BoundedCoalescent_backward_sampler_c", (DL_FUNC) &_BoundedCoalescent_backward_sampler_c, 6},
     {"_BoundedCoalescent_block_coalescences_c", (DL_FUNC) &_BoundedCoalescent_block_coalescences_c, 4},
     {"_BoundedCoalescent_separate_coalescences_c", (DL_FUNC) &_BoundedCoalescent_separate_coalescences_c, 5},
     {"_BoundedCoalescent_constrain_coalescences_c", (DL_FUNC) &_BoundedCoalescent_constrain_coalescences_c, 5},
     {"_BoundedCoalescent_sample_coalescence_time_c", (DL_FUNC) &_BoundedCoalescent_sample_coalescence_time_c, 4},
-    {"_BoundedCoalescent_forward_algorithm_c", (DL_FUNC) &_BoundedCoalescent_forward_algorithm_c, 4},
-    {"_BoundedCoalescent_sample_bounded_times_c", (DL_FUNC) &_BoundedCoalescent_sample_bounded_times_c, 5},
     {"_BoundedCoalescent_homochronous_probability", (DL_FUNC) &_BoundedCoalescent_homochronous_probability, 4},
+    {"_BoundedCoalescent_forward_algorithm_c", (DL_FUNC) &_BoundedCoalescent_forward_algorithm_c, 5},
+    {"_BoundedCoalescent_backward_sampler_c", (DL_FUNC) &_BoundedCoalescent_backward_sampler_c, 6},
+    {"_BoundedCoalescent_sample_bounded_times_c", (DL_FUNC) &_BoundedCoalescent_sample_bounded_times_c, 5},
     {"_BoundedCoalescent_bounded_times_likelihood_c", (DL_FUNC) &_BoundedCoalescent_bounded_times_likelihood_c, 5},
     {"_BoundedCoalescent_rejection_bounded_times", (DL_FUNC) &_BoundedCoalescent_rejection_bounded_times, 5},
     {"_BoundedCoalescent_sample_topology_c", (DL_FUNC) &_BoundedCoalescent_sample_topology_c, 3},
